refactor: presentation 패키지 구성 변경
- 화면 단위로 presentation 분리
