+++ app/src/androidTest/java/woowacourse/movie/view/ReservationResultActivityTest.kt
@@ -11,6 +11,7 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import org.junit.Rule
 import org.junit.Test
 import woowacourse.movie.R
+import woowacourse.movie.presentation.view.reservation.result.ReservationResultActivity
 
 class ReservationResultActivityTest {
     private val testContext = ApplicationProvider.getApplicationContext<Context>()
