+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.view.SeatSelectionActivity"
+            android:name=".presentation.view.reservation.seat.SeatSelectionActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.view.ReservationResultActivity"
+            android:name=".presentation.view.reservation.result.ReservationResultActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.view.MovieDetailActivity"
+            android:name=".presentation.view.reservation.detail.MovieDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.view.MainActivity"
+            android:name=".presentation.view.screening.ScreeningActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
