+++ app/src/main/java/woowacourse/movie/presentation/view/screening/adapter/MovieViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.presentation.adapter
+package woowacourse.movie.presentation.view.screening.adapter
 
 import android.view.View
 import android.widget.ImageView
