+++ app/src/main/res/layout/activity_reservation_result.xml
@@ -5,7 +5,7 @@
     android:id="@+id/reservation_result_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.view.ReservationResultActivity">
+    tools:context=".presentation.view.reservation.result.ReservationResultActivity">
 
     <TextView
         android:id="@+id/reservation_cancel_info"
