+++ app/src/main/res/layout/activity_screening.xml
@@ -8,13 +8,13 @@
 
         <variable
             name="main"
-            type="woowacourse.movie.presentation.view.MainActivity" />
+            type="woowacourse.movie.presentation.view.screening.ScreeningActivity" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.view.MainActivity">
+        tools:context=".presentation.view.screening.ScreeningActivity">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/movieList"
