+++ app/src/main/res/layout/activity_seat_selection.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.view.SeatSelectionActivity">
+    tools:context=".presentation.view.reservation.seat.SeatSelectionActivity">
 
     <ScrollView
         android:layout_width="match_parent"
